refactor: relocate KgPythonDictFramework into compute_frameworks/
The package's only compute framework lived under feature_groups/kg/ while
the dedicated compute_frameworks/ package was an empty docstring-only
namespace, so a contributor looking for compute frameworks found nothing
and had to discover the framework in the feature-groups tree (issue #9).

Move python_dict_kg_framework.py (and its direct unit test) into
compute_frameworks/, update every import (base.py, test helpers, demos,
README), and replace the empty compute_frameworks/__init__.py docstring
with a one-liner naming the framework and the FG base that pins it.

The framework only depends on mloda's PythonDictFramework, so the move
introduces no circular import and gives a clean dependency direction
(feature_groups -> compute_frameworks).

Closes #9
