+++ demo/demo_kg_connectors.py
@@ -67,7 +67,7 @@ def helpers():
         Feature as _Feature,
         mloda as _mloda,
     )
-    from open_kgo.feature_groups.kg.python_dict_kg_framework import (
+    from open_kgo.compute_frameworks.python_dict_kg_framework import (
         KgPythonDictFramework as _KgPythonDictFramework,
     )
